challenge/requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==5.1.1
Django==4.1.2
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
django-user-agents==0.4.0
djangorestframework==3.14.0
executing==1.1.1
idna==3.4
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pytz==2022.4
requests==2.28.1
screen==1.0.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
traitlets==5.4.0
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
wcwidth==0.2.5
xlwt==1.3.0
